+++ xarray/core/dataset.py
@@ -53,12 +53,12 @@ from . import (
 from .alignment import _broadcast_helper, _get_broadcast_dims_map_common_coords, align
 from .arithmetic import DatasetArithmetic
 from .common import DataWithCoords, _contains_datetime_like_objects
+from .computation import unify_chunks
 from .coordinates import (
     DatasetCoordinates,
     assert_coordinate_consistent,
     remap_label_indexers,
 )
-from .dask_array_ops import unify_chunks
 from .duck_array_ops import datetime_to_numeric
 from .indexes import (
     Index,
